Added a failsafe for checking that the repos deletion path is sane

+++ lib/git_backend.rb
@@ -20,7 +20,11 @@ class GitBackend
     end
     
     def delete!(repos_path)
-      FileUtils.rm_rf(repos_path)
+      if repos_path.index(GitoriousConfig["repository_base_path"]) == 0
+        FileUtils.rm_rf(repos_path)
+      else
+        raise "bad path"
+      end
     end
     
     def repository_has_commits?(repos_path)
